% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vancomycin_Dose_Calculator.R
\name{Vanco_SS_dose_calculator}
\alias{Vanco_SS_dose_calculator}
\title{Vancomycin Dose Calculator}
\usage{
Vanco_SS_dose_calculator(
  pre_level,
  post_level,
  dose,
  regimen,
  infusion_time,
  peak_time,
  trough_time,
  peak_conc,
  trough_conc
)
}
\arguments{
\item{pre_level}{Patient level taken prior to providing the next dose (mg/L)}

\item{post_level}{Patient level taken after infusion time of dose (mg/L)}

\item{dose}{Amount of drug given to the patient (mg)}

\item{regimen}{Dosing interval currently used for the patient (Q8h, Q12h, Q18h, Q24h, Q36h, Q48h)}

\item{infusion_time}{Time over which the drug is infused (h)}

\item{peak_time}{The time when the post level is drawn (h)}

\item{trough_time}{The time with the pre_level is drawn (h)}

\item{peak_conc}{The ideal peak concentration for the patient (mg/L)}

\item{trough_conc}{The ideal trough concentration for the patient (mg/L)}
}
\value{
The recommended dosing interval is returned through this function
}
\description{
The function calculates a new dosing regimen for patients receiving Vancomycin using the Sawchuk-Zaske calculation
strategy. This strategy determines the elimination rate constant(k), the half-life(t1/2), the volume of distribution (Vd), the
new dosing interval (tau), the new dosing rate recommendation (k0), the expected peak level (C 1h), and the expected trough
level (C tr). Any relevant equations can be found online or by viewing the equations in the function. Two or three serum
concentrations are needed, along with the assumption that the patient is at steady-state (receiving their 4th or 5th dose
depending on the half-life of the drug). Using the patient parameters, individualized dosing regimens can be created to achieve
ideal target levels. The goal of this function is to facilitate the dosing adjustments that are typically conducted in a
hospital setting, and to allow clinical pharmacists to make patient specific judgements quicker.
}
\examples{

Vanco_SS_dose_calculator(11.0, 22.1, 1000, 12, 1.5, 1.33333, 55/60, 35, 17.5) # Should be "This function recommends a new dose of", 1500,"given over Q",12,"h",". This will produce a peak of", 33.15,"and a trough value of", 15.2692370301831
Vanco_SS_dose_calculator(25.0, 33.5, 1500, 8, 1.5, 115/60, 92/60, 33.5, 17.5) # Should be "This function recommends a new dose of", 1500,"given over Q",8,"h",". This will produce a peak of", 33.5,"and a trough value of", 21.5794478677179
Vanco_SS_dose_calculator(c(25,25), c(33.5, 43.0), c(1500,1000), c(8,8), c(1.5,1), c(115/60,1.5), c(92/60, 1), c(33.5,35), c(17.5:17.5)) # Should be "This function recommends a new dose of", 1500, 750,"given over Q",8,8,"h",". This will produce a peak of", 33.5, 32.25,"and a trough value of", 21.5794478677179, 16.6211709853472

}
